+++ tests/test_39_plus.py
@@ -3,7 +3,7 @@ from typing import Annotated
 import pytest
 
 from koda_validate import MinLength, StringValidator
-from koda_validate.decorators import (
+from koda_validate.signature import (
     InvalidArgsError,
     InvalidReturnError,
     validate_signature,
